single-source multi-threading tests
